requirements.txt
httpx>=0.27
python-dotenv>=1.0
openai>=1.42.0
psycopg2-binary==2.9.9
fastapi>=0.111
uvicorn>=0.30
jinja2>=3.1
sse-starlette>=1.8.2
requests>=2.32.0
SQLAlchemy>=2.0.0
PyYAML>=6.0.0
google-cloud-storage>=2.16.0
google-api-python-client>=2.130
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
llama-index>=0.13.6
llama-index-embeddings-openai>=0.2.0
llama-index-vector-stores-postgres>=0.6.0
rank-bm25==0.2.2
pypdf>=4.3.1
python-docx>=1.1.2
pytesseract>=0.3.13
Pillow>=10.0.0
PyMuPDF>=1.24.10
duckduckgo-search>=6.2.6
trafilatura>=1.7.0
tldextract>=5.1.2
# Firecrawl (optional). If missing, we fall back to trafilatura
firecrawl>=1.3.0
beautifulsoup4>=4.12.3
mangum>=0.17.0
